+++ RELEASE_NOTES.md
@@ -25,6 +25,7 @@ The reason for this change is that it's easier to find, than to have standard st
 * `Keyboard.ButtonStyle` has a `standard(for:action:isPressed:)` builder.
 * `Keyboard.ButtonBorderStyle` has a new `standard(for:action:)` builder.
 * `Keyboard.ButtonShadowStyle` has a new `standard(for:action:)` builder.
+* `KeyboardFont` has new `standard` font, font sizeit   and font weight builders.
 
 ### 🗑️ Deprecations